+++ spec/ruby/core/enumerable/grep_v_spec.rb
@@ -20,15 +20,25 @@ describe "Enumerable#grep_v" do
     $&.should == "e"
   end
 
-  it "sets $~ to the last match when given no block" do
-    "z" =~ /z/ # Reset $~
-    ["abc", "def"].grep_v(/e/).should == ["abc"]
+  ruby_version_is ""..."3.0.0" do
+    it "sets $~ to the last match when given no block" do
+      "z" =~ /z/ # Reset $~
+      ["abc", "def"].grep_v(/e/).should == ["abc"]
 
-    # Set by the match of "def"
-    $&.should == "e"
+      # Set by the match of "def"
+      $&.should == "e"
 
-    ["abc", "def"].grep_v(/b/)
-    $&.should == nil
+      ["abc", "def"].grep_v(/b/)
+      $&.should == nil
+    end
+  end
+
+  ruby_version_is "3.0.0" do
+    it "does not set $~ when given no block" do
+      "z" =~ /z/ # Reset $~
+      ["abc", "def"].grep_v(/e/).should == ["abc"]
+      $&.should == "z"
+    end
   end
 
   describe "without block" do