+++ NEWS.md
@@ -371,7 +371,7 @@ Outstanding ones only.
 
 * Bundler
 
-    * Update to Bundler 2.2.1
+    * Update to Bundler 2.2.2
 
 * CSV
 
@@ -434,7 +434,7 @@ Outstanding ones only.
 
 * RubyGems
 
-    * Update to RubyGems 3.2.1
+    * Update to RubyGems 3.2.2
 
 ## Compatibility issues